backend/requirements.txt
fastapi
uvicorn[standard]
GitPython
openai
